requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==2.15.8
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.7
dj-database-url==2.2.0
Django==3.2.25
django-cors-headers==4.1.0
django-filter==23.5
django-stubs==4.2.0
django-stubs-ext==4.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
ez-setup==0.9
gunicorn==23.0.0
gunicorn==23.0.0
idna==3.10
importlib-metadata==4.2.0
isort==5.11.5
kombu==5.2.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
numpy==1.21.6
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
platformdirs==4.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.8.0
pylint==2.17.7
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.0.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==11.1.0
tomli==2.0.1
tomlkit==0.12.5
typed-ast==1.5.5
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.7.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
zipp==3.15.0
